Switch to a custom SInt type.
The one in "snumber" was meant for different purposes, and trying to add
arithmetic to it seems somewhat risky.  With a custom Int-only type, we
can know concretely what the arithmetic functions do w.r.t. overflow,
etc.

SNumber also works with Kinds.Integer.Integer instead of Nat, which
makes it a bit less convenient -- for example, we had to wrap Nats in
Pos in many places, and would have to do our own sign checks when
subtracting them.

It's a bit aesthetically unsatisfying to have two similar
implementations, but it's been so much nicer in practice to use the
far-simpler SInt here that I think it's worth the redundancy.

Change-Id: If2d573e5fff26004f0428c5fc5dd112a345a2ec1