Remove the Num instance of Fin.
Since most of the Num methods for Fin would be partial, and Num doesn't
have HasCallStack constraints on its methods, having a Num instance is a
major source of impossible-to-debug runtime errors: if you use a literal
too large for the Fin type at runtime, or evaluate subtraction resulting
in a negative number, you get nothing but the name of the Num method it
came from (and not the location of the call site), with no way to narrow
it down within your entire codebase, except by making a profiling build
and reproducing the failure with stack trace profiling enabled.

Instead, we provide the corresponding functionality with
operators/functions with distinct names and `HasCallStack` constraints:
`fin` for constructing `Fin`s and `+!` etc. for arithmetic.  This way
partial arithmetic expressions are visually distinguished, and you
always get the location of the problematic call site when it fails.

To recover some of the convenience of being able to use integer
literals/patterns for `Fin`s, check out the
[https://www.github.com/google/hs-dependent-literals](dependent-literals)
plugin.

Change-Id: I2d8e7ab3d921d7d78511494408fbd5d7c7c88549