Use `SInt` internally for `Fin` bound values.
When passing the runtime value of the bound as `KnownNat`, we incur a
runtime conversion from `Natural` to `Int`, unless we get lucky and
inlining makes it go away (e.g. in simple cases with only
statically-known bounds).  Instead, if we pass the value as `SInt`, it's
already represented as an `Int`, and there's no conversion to do at
runtime.

The more-convenient `KnownNat`-based API is now a thin wrapper over the
explicit-bounds-passing one.

Change-Id: I53c8493896147364ec89e8d7dacd5bbf9a38b4a2